push bug

+++ repositories/ratingRepo.go
@@ -10,7 +10,7 @@ type RatingRepo interface {
 	AddReview(review *entities.Review) error
 	GetReviewByUserAndProduct(userID, productID int) (*entities.Review, error)
 	CountReviewsByProduct(productID int) (int, error)
-	SumRatingByProduct(productID int) (int, error)
+	SumRatingByProduct(productID int) (float64, error) // Perubahan tipe return menjadi float64
 }
 
 type ratingRepo struct {
@@ -52,8 +52,8 @@ func (r *ratingRepo) CountReviewsByProduct(productID int) (int, error) {
 	return int(count), nil
 }
 
-func (r *ratingRepo) SumRatingByProduct(productID int) (int, error) {
-	var totalRating int
+func (r *ratingRepo) SumRatingByProduct(productID int) (float64, error) { // Perubahan tipe return menjadi float64
+	var totalRating float64 // Ubah menjadi float64
 	err := r.db.Model(&entities.Review{}).Where("product_id = ?", productID).Select("SUM(rating)").Scan(&totalRating).Error
 	if err != nil {
 		return 0, err
