+++ services/detailprodukServices.go
@@ -28,19 +28,22 @@ func (s *productDetailServices) LihatProductByID(ProductID int) (*entities.Produ
 }
 
 func (s *productDetailServices) UpdateProductRating(productID int) error {
+	// Mendapatkan jumlah ulasan untuk produk
 	totalReviews, err := s.reviewRepo.CountReviewsByProduct(productID)
 	if err != nil {
 		return err
 	}
 
+	// Mendapatkan jumlah rating total untuk produk
 	totalRating, err := s.reviewRepo.SumRatingByProduct(productID)
 	if err != nil {
 		return err
 	}
 
+	// Menghitung rating baru dengan pembagian total rating / total ulasan
 	newRating := 0.0
 	if totalReviews > 0 {
-		newRating = float64(totalRating) / float64(totalReviews)
+		newRating = totalRating / float64(totalReviews) // totalRating sekarang bertipe float64
 	}
 
 	err = s.productRepo.UpdateProductRating(productID, newRating)
@@ -48,6 +51,7 @@ func (s *productDetailServices) UpdateProductRating(productID int) error {
 		return err
 	}
 
+	// Memperbarui jumlah ulasan produk di repository
 	err = s.productRepo.UpdateTotalReview(productID, int(totalReviews))
 	return err
 }
